pyproject.toml
[tool.poetry]
name = "edubook-python"
version = "0.1.0"
description = ""
authors = ["pybsh <jnhyeon18@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

